pyproject.toml
[project]
name = "m9-ticket-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "snowflake>=1.12.0",
    "sqlmodel>=0.0.37",
    "upstash-redis>=1.6.0",
]
